host info update
Deals with Issue #3.
Now displays today's date in a mm/dd/yyyy structure to make it look pretty since the normal Date() is kinda boring. I researched on stack overflow to find stuff you can do with dates to implement this.
Now displays Status that can be update via user input. ALSO, I was looking at some past OS projects and noticed the use of default statuses that I thought was a cool touch, so implemented that as well.