figured out upperCase status
Issue #3 resolved
Realized that the buffer is automatically set to lowerCase before getting split, so instead I moved it to only lower case the first element after the split so that the args will now be able to be mixed case. This actually helps make my stairs shell command a little better as well which is nice.