Memory Table
HTML/TS work to create a small scrolling memory table with initialized values. Took me a while to realize why my starting value of each row was missing the padded zeroes. Turns out .toString(16) doesn't do the padding zeores for you, it only changes it to hex. So to fix that I added a while loop to add zeroes until the string had a length of 4.