CPU Table
Created the CPU table in HTML and initialized the values in Control.ts. Very similar to what I did with the memory table. Included the IR column because I saw it on the rhinOS project and I thought it was neat and cool for showing which instruction is currently being fetched, decoded, and executed.