PCB table & CPU instructions
PCB table was implemented the exact same way as the CPU table.
The CPU instructions were copied straight from my Org & Arch project but removed the interrupt step since in lecture we were told it is unneeded. I also removed some of the extra op codes.
Moving forward I need to create a memory class so that the cpu can access it as well as an mmu class.