Finally seeing progress
I think what makes this project so frustrating to me is that it took forever to see any progress at all. I did a ton of coding but to me it feels like I did nothing unless I see it with my own eyes so I had to make sure the display shows the memory correctly before I moved on. It also took me forever playing around with string methods to get every two characters in the user program input. Even before that understanding and being able to implement the cpu, ma, mm, and pcb took a while but after doing a lot of research and trying to separate my Org & Arch class understanding a little, I think it is better. I basically separated my MMU from Org & Arch to my MM and MA. For the PCB, it loads the memory and runs the CPU based on the memory loaded. I know in the future there will be multiple PCBs but for now just leaving it as one that can only have pid = 0 was easier for now. The specific hints on the IP2 also helped a lot. This should handle all of the implementation requirements unless the PCB is too underdeveloped. issues #16-#20

I believe shell command load is now should be implemented correctly (at least for IP2 since I didn't allow for more than one pcb block yet) which handles issue #9 .

I also created the shell run command which should be okay which handles issue #10. It can only run pid 0 right now.

The CPU has A9 and 00 just to play with run and see if it is displaying correctly.

Looking Forward:
Hoping I can get the op codes to work correctly so I can hand in the project in time. I need little endian conversion too which shouldn't be too bad since I already have the method from Org & Arch.
I also need to make sure I update the PCB with the CPU.