Running 3 programs at once
Working on the scheduler and dispatcher issue #30 and implementing the context switch.
While working on this I noticed that my scheduler actually needed to be modified because it was never calling the interrupt. I looked over the diagram and noticed that a context switch is actually called before the cpu executes so I removed the code that tells the cpu to execute in runall and checked that if there is no pcb in the cpu call the context switch because we know that it's the first pcb if there was no pcb before it.
Also while working on the context switch I realized that when I was enqueuing the last pcb and dequeuing the next pcb it would cause my output to be weird/wrong when I performed runall. For example on the a0a1a2adone program it would just do a0adone. I then realized the problem was that the cpu had no clue where to continue off of when a new program get context switched so I added a new method that will change all the cpu variables like the program counter to be the same as what the pcb being dequeued has.
Now 3 programs can be ran at once!
Honestly I'm a little dumb for procrastinating the ready queue display till the end since it probably could have helped me get to this solution much sooner but I kinda hate html tables.
 Also one thing that sucks is that my single step doesn't work anymore so it will probably have to be modified.