Read and Write Progress
A lot of progress on read and write but there still needs more error checking and access to bigger dataStrings. But simple tests should work correctly like "hello". Need to still make it work for dataStrings with spaces though. Also made some side methods that should be helpful for the diskSystem like finding the key that contains the fileName and appending commas for using next TSB keys as current TSB keys.