Finally got 4 12Done programs to somewhat work
#43
Oh my god that took forever to finally get the right output but I still have a ton of issues. The only way I was able to finish all processes was to never stop running the pc or else the last program on disk will be left behind. I also had a ton of issues with rollOut setting data from memory as just zeroes so I simplified it by putting the entire memory string into one getItem which is technically not correct. This is what I get for not starting earlier to be fair